requirements.txt
google-generativeai>=0.8.0
pandas>=2.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
seaborn>=0.12.0
